+++ event-store-metric-meters/src/main/java/uk/gov/justice/services/eventstore/metrics/tags/TagNames.java
@@ -4,7 +4,9 @@ public enum TagNames {
 
 
     SOURCE_TAG_NAME("source"),
-    COMPONENT_TAG_NAME("component");
+    COMPONENT_TAG_NAME("component"),
+    SERVICE_TAG_NAME("service"),
+    ENV_TAG_NAME("env");
 
     private final String tagName;