+++ CHANGELOG.md
@@ -7,9 +7,11 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 ### Added
 - New table stream_statistic to include statistics of streams per source, component and state
 - Added StreamStatisticTimerBean to fill stream_statistic every configured milliseconds
+- Global tags for micrometer gauges
 ### Changed
 - Add framework rest endpoint to fetch streams by errorHash
 - Metrics lookup is now done using `source` and `component`
+- Individual Gauges refactored to Factory
 
 # [17.103.0-M13] - 2025-06-18
 ### Added